WEBDEV-8951 Move maskedIcon into src/util so components can share it
`maskedIcon` arrived with the item navigator in #64, but nothing about it is
navigator-specific: it renders a standalone `.svg` as a CSS-masked span so the
glyph stays recolorable and its path data stays out of the JS bundle, and it
carries the quoted-`url()` fix that keeps bundler-inlined `data:` URIs from
silently dropping the whole `mask-image` declaration.

The status indicator needs exactly that, so the helper and its test move to
`src/util/` rather than a second copy carrying the same fix. Only import paths
change for the navigator; the helper itself is untouched.
