+++ src/util/masked-icon.test.ts
