+++ src/util/masked-icon.ts
